requirements.txt
# -------------------------
# Core deep learning
# -------------------------
torch>=1.13.0
torchvision>=0.14.0

# -------------------------
# Scientific computing
# -------------------------
numpy>=1.21
scipy>=1.7

# -------------------------
# Medical / volume I/O
# -------------------------
tifffile>=2022.5.4

# -------------------------
# Configuration
# -------------------------
PyYAML>=6.0

# -------------------------
# Logging & visualization
# -------------------------
tensorboard>=2.10

# -------------------------
# Utilities
# -------------------------
tqdm>=4.64
